chatbot/openai/requirements.txt
requests==2.32.5
PyYAML==6.0.3
termcolor==3.1.0
emoji==2.15.0
transformers==4.57.3
torch==2.2.2
numpy==1.26.4
langchain==0.3.27
langchain-community==0.3.27
jsonlines==4.0.0
tiktoken==0.12.0
openai==2.13.0
chromadb==1.3.7
rich==14.2.0